+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 2.4.5 (unreleased)
 ------------------
 
-- Return bookings in ascending date order so the ones for today are shown as first
-  [pnicolli]
+- sort_on/sort_order in restapi bookings and xlsx
+  [mamico]
 
 
 2.4.4 (2024-01-25)